requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
langchain-openai>=0.2.0
chromadb>=0.5.0
streamlit>=1.40.0
streamlit-chat>=0.1.1
python-dotenv>=1.0.0
pypdf>=5.0.0
pydantic>=2.10.0
tiktoken>=0.7.0
